Moved datepicker range to ssr frindly arhitecture
