+++ src/ui.form-date-picker-range/UDatePickerRangePeriodMenu.vue
@@ -209,11 +209,11 @@ import {
   getYearDateList,
   getQuartersDateList,
   getMonthsDateList,
-} from "../services/dateRange.service";
+} from "./utilDateRange.js";
 
-import UButton from "../../ui.button";
+import UButton from "../ui.button/UButton.vue";
 
-import { PERIOD } from "../constants";
+import { PERIOD } from "./constants.js";
 
 const props = defineProps({
   locale: {
